+++ docs/documentation/docs/about/release-notes.md
@@ -4,7 +4,8 @@
 
 **Enhancements**
 
-- PropertyFieldCollectionData: Hide save button when "Add and save" is shown [#88](https://github.com/SharePoint/sp-dev-fx-property-controls/issues/88)
+- `PropertyFieldCollectionData`: override placeholder for the inputs [#87](https://github.com/SharePoint/sp-dev-fx-property-controls/issues/87)
+- `PropertyFieldCollectionData`: Hide save button when "Add and save" is shown [#88](https://github.com/SharePoint/sp-dev-fx-property-controls/issues/88)
 
 ## 1.8.0